+++ BillApp.Web/Models/DashboardViewModels.cs
@@ -8,8 +8,8 @@ namespace BillApp.Web.Models
     public class DashboardViewModels
     {
         public int NumInvoices { get; set; }
-        public float TotalInvoices { get; set; }
+        public double TotalInvoices { get; set; }
         public int NumCustomers { get; set; }
-        public float NumOfArticles { get; set; }
+        public double NumOfArticles { get; set; }
     }
 }
